backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
flask-socketio==5.3.6
python-socketio==5.11.1
python-dotenv==1.0.1
mistralai==1.12.4
elevenlabs==1.50.5
wandb==0.19.8
gunicorn==23.0.0
requests==2.32.3
firebase-admin==6.5.0
google-generativeai==0.8.3
Pillow==11.0.0
